show the listpost and restructure of folder and design
