+++ assets/css/post.css
@@ -1,0 +1,23 @@
+.wrapper-list {
+    padding-bottom: 10px;
+    border-bottom: 0.8px solid black;
+
+    width: 100%;
+    display: flex;
+    flex-flow: row nowrap;
+}
+.post-info {
+    width: 80%;
+    display: flex;
+    flex-direction: column;
+}
+.post-list {
+    width: 90%;
+}
+.post-buttons {
+    width: 10%;
+    display: flex;
+    flex-flow: row nowrap;
+    align-items: flex-end;
+    padding: 0 0 10px 10px;
+}
