+++ src/app.js
@@ -1,3 +1,7 @@
+import { getPostApi } from "./services/api_handler.js";
+
+getPostApi()
+
 //-------------button functionalities----------------------//
 let imgs = document.querySelectorAll("img");
 
@@ -22,32 +26,3 @@ function closePost(){
 function deletePost(){
     console.log("Deleting post")
 }
-
-//------------------modal rendering-------------------------//
-
-let posts = document.querySelectorAll(".post")
-let modal = document.getElementById("modal")
-
-posts.forEach((p) => {
-    p.addEventListener("click", renderModal)
-})
-
-function renderModal(){
-    modal.style.display = "block"
-}
-
-
-//----------------closing modal--------------------------------//
-
-let closeBtn = document.getElementById("close-modal")
-
-closeBtn.addEventListener("click", () =>{
-    modal.style.display = "none";
-})
-
-/*document.addEventListener("click", (e) =>{
-    console.log(e.target)
-    let modalChildren = modal.children;
-    console.log(modalChildren)                -------------> cerrar modal clickando fuera
-})*/
-
