+++ src/services/api_handler.js
@@ -1,21 +1,22 @@
+import { postListView } from "../view/listPost.js";
 
-function getPostApi() {
+export function getPostApi() {
     return $.ajax("https://jsonplaceholder.typicode.com/posts", {
         method: 'GET'
     }).then(
         function success(json, statusText, jqXHR) {
-            json.forEach(post => {
-                // console.log('post', post);
-            });
-            console.log('statusText success', statusText);
+            json.forEach((post) => {
+                postListView(post)
+            })
+            // console.log('statusText success', statusText);
         },
         function failed(jqXHR, errorStatus, errorMessage) {
-            console.log(errorMessage);
+            // console.log(errorMessage);
         }
     )
 }
 
-function getUserApi() {
+export function getUserApi() {
     return $.ajax("https://jsonplaceholder.typicode.com/users", {
     method: 'GET'
     }).then(
@@ -31,13 +32,13 @@ function getUserApi() {
     )
 }
 
-function getCommentsApi() {
+export function getCommentsApi() {
     return $.ajax("https://jsonplaceholder.typicode.com/comments", {
     method: 'GET'
     }).then(
         function success(json, statusText, jqXHR) {
             json.forEach(comment => {
-                console.log('comment', comment);
+                // console.log('comment', comment);
             });
             console.log('statusText success', statusText);
         },
@@ -46,7 +47,3 @@ function getCommentsApi() {
         }
     )
 }
-
-getPostApi()
-getUserApi()
-getCommentsApi()
